Variety of performance improvements.
