+++ public/css/tags.css
@@ -1,58 +1,14 @@
-.tags-input-disabled {
-	border: 1px solid #CED4DA;
-	display: inline-block;
-  border-radius: 6px;
-	width: 100%;
-}
-
-.tags-input-disabled .tag {
-	font-size: 85%;
-	padding: 0.5em .75em;
-	margin: .25em .1em;
-	display: inline-block;
-	background-color: #E9ECEF;
-	border: 1px solid #DEE2E6;
-	transition: all .1s linear;
-}
-
-.tags-input {
-	position: relative;
-	display:block;
-	border: 1px solid #CED4DA;
-	background-color: white;
-  border-radius: 6px;
-	cursor: text;
-}
-
 .tag {
 	background-color: #E9ECEF;
 }
 
-.tags-input .tag, .autocard-tags .tag, .tag-item {
+.autocard-tags .tag, .tag-item {
 	font-size: 85%;
 	padding: 0.5em .75em;
 	margin: .25em .1em;
 	display: inline-block;
 	border: 1px solid #DEE2E6;
 	transition: all .1s linear;
-}
-
-.tags-input .tag .close-tag::after {
-	content: '×';
-	font-weight: bold;
-	display: inline-block;
-	transform: scale(1.4);
-	margin-left: .75em;
-}
-.tags-input .tag .close-tag:hover::after {
-	color:red;
-}
-
-.tags-input .main-input {
-	border: 0;
-	outline: 0;
-	background: 0;
-	padding: .5em .2em;
 }
 
 .autocard-tags {
