--- public/js/editcube.js
@@ -1,329 +1,0 @@
-var comparing = false;
-if ($('#in_both').length) {
-  comparing = true;
-  var in_both = JSON.parse($('#in_both').val());
-  var only_a = JSON.parse($('#only_a').val());
-  var only_b = JSON.parse($('#only_b').val());
-  view = 'table';
-}
-
-var cube = reactProps.cards || JSON.parse($('#cuberaw').val());
-cube.forEach(function(card, index) {
-  card.index = index;
-});
-
-function GetColorIdentity(colors) {
-  if (colors.length == 0) {
-    return 'Colorless';
-  } else if (colors.length > 1) {
-    return 'Multicolored';
-  } else if (colors.length == 1) {
-    switch (colors[0]) {
-      case 'W':
-        return 'White';
-        break;
-      case 'U':
-        return 'Blue';
-        break;
-      case 'B':
-        return 'Black';
-        break;
-      case 'R':
-        return 'Red';
-        break;
-      case 'G':
-        return 'Green';
-        break;
-      case 'C':
-        return 'Colorless';
-        break;
-    }
-  }
-}
-
-function getSorts() {
-  return [
-    'Artist',
-    'CMC',
-    'Color Category',
-    'Color Count',
-    'Color Identity',
-    'Color',
-    'Date Added',
-    'Finish',
-    'Guilds',
-    'Legality',
-    'Loyalty',
-    'Manacost Type',
-    'Power',
-    'Price',
-    'Price Foil',
-    'Rarity',
-    'Set',
-    'Shards / Wedges',
-    'Status',
-    'Subtype',
-    'Supertype',
-    'Tags',
-    'Toughness',
-    'Type',
-    'Types-Multicolor',
-    'Unsorted',
-  ];
-}
-
-function getLabels(sort) {
-  if (sort == 'Color Category') {
-    return ['White', 'Blue', 'Black', 'Red', 'Green', 'Multicolored', 'Colorless', 'Lands'];
-  } else if (sort == 'Color Identity') {
-    return ['White', 'Blue', 'Black', 'Red', 'Green', 'Multicolored', 'Colorless'];
-  } else if (sort == 'CMC') {
-    return ['0', '1', '2', '3', '4', '5', '6', '7', '8+'];
-  } else if (sort == 'CMC2') {
-    return ['0-1', '2', '3', '4', '5', '6', '7+'];
-  } else if (sort == 'CMC-Full') {
-    // All CMCs from 0-16, with halves included, plus Gleemax at 1,000,000.
-    return Array.from(Array(33).keys())
-      .map((x) => (x / 2).toString())
-      .concat(['1000000']);
-  } else if (sort == 'Color') {
-    return ['White', 'Blue', 'Black', 'Red', 'Green', 'Colorless'];
-  } else if (sort == 'Type') {
-    return [
-      'Creature',
-      'Planeswalker',
-      'Instant',
-      'Sorcery',
-      'Artifact',
-      'Enchantment',
-      'Conspiracy',
-      'Contraption',
-      'Phenomenon',
-      'Plane',
-      'Scheme',
-      'Vanguard',
-      'Land',
-      'Other',
-    ];
-  } else if (sort == 'Supertype') {
-    return ['Snow', 'Legendary', 'Tribal', 'Basic', 'Elite', 'Host', 'Ongoing', 'World'];
-  } else if (sort == 'Tags') {
-    var tags = [];
-    cube.forEach(function(card, index) {
-      card.tags.forEach(function(tag, index2) {
-        if (tag.length > 0 && !tags.includes(tag)) {
-          tags.push(tag);
-        }
-      });
-    });
-    return tags.sort();
-  } else if (sort == 'Date Added') {
-    var days = [],
-      formattedDay;
-    cube.forEach(function(card, index) {
-      formattedDay = ISODateToYYYYMMDD(card.addedTmsp);
-      if (formattedDay === undefined) {
-        formattedDay = 'unknown';
-      }
-      if (!days.includes(formattedDay)) {
-        days.push(formattedDay);
-      }
-    });
-    return days.sort();
-  } else if (sort == 'Status') {
-    return ['Not Owned', 'Ordered', 'Owned', 'Premium Owned'];
-  } else if (sort == 'Finish') {
-    return ['Non-foil', 'Foil'];
-  } else if (sort == 'Guilds') {
-    return ['Azorius', 'Dimir', 'Rakdos', 'Gruul', 'Selesnya', 'Orzhov', 'Golgari', 'Simic', 'Izzet', 'Boros'];
-  } else if (sort == 'Shards / Wedges') {
-    return ['Bant', 'Esper', 'Grixis', 'Jund', 'Naya', 'Abzan', 'Jeskai', 'Sultai', 'Mardu', 'Temur'];
-  } else if (sort == 'Color Count') {
-    return ['0', '1', '2', '3', '4', '5'];
-  } else if (sort == 'Set') {
-    var sets = [];
-    cube.forEach(function(card, index) {
-      if (!sets.includes(card.details.set.toUpperCase())) {
-        sets.push(card.details.set.toUpperCase());
-      }
-    });
-    return sets.sort();
-  } else if (sort == 'Artist') {
-    var artists = [];
-    cube.forEach(function(card, index) {
-      if (!artists.includes(card.details.artist)) {
-        artists.push(card.details.artist);
-      }
-    });
-    return artists.sort();
-  } else if (sort == 'Rarity') {
-    return ['Common', 'Uncommon', 'Rare', 'Mythic'];
-  } else if (sort == 'Unsorted') {
-    return ['All'];
-  } else if (sort == 'Subtype') {
-    var types = [];
-    cube.forEach(function(card, index) {
-      if (card.type_line.includes('—')) {
-        var subtypes = card.type_line.substr(card.type_line.indexOf('—') + 1).split(' ');
-        subtypes.forEach(function(subtype, index) {
-          if (!types.includes(subtype.trim()) && subtype.trim().length > 0) {
-            types.push(subtype.trim());
-          }
-        });
-      }
-    });
-    return types.sort();
-  } else if (sort == 'Types-Multicolor') {
-    return [
-      'Creature',
-      'Planeswalker',
-      'Instant',
-      'Sorcery',
-      'Artifact',
-      'Enchantment',
-      'Conspiracy',
-      'Contraption',
-      'Phenomenon',
-      'Plane',
-      'Scheme',
-      'Vanguard',
-      'Azorius',
-      'Dimir',
-      'Rakdos',
-      'Gruul',
-      'Selesnya',
-      'Orzhov',
-      'Golgari',
-      'Simic',
-      'Izzet',
-      'Boros',
-      'Bant',
-      'Esper',
-      'Grixis',
-      'Jund',
-      'Naya',
-      'Abzan',
-      'Jeskai',
-      'Sultai',
-      'Mardu',
-      'Temur',
-      'Non-White',
-      'Non-Blue',
-      'Non-Black',
-      'Non-Red',
-      'Non-Green',
-      'Five Color',
-      'Land',
-      'Other',
-    ];
-  } else if (sort == 'Legality') {
-    return ['Standard', 'Modern', 'Legacy', 'Vintage', 'Pauper'];
-  } else if (sort == 'Power') {
-    var items = [];
-    cube.forEach(function(card, index) {
-      if (card.details.power) {
-        if (!items.includes(card.details.power)) {
-          items.push(card.details.power);
-        }
-      }
-    });
-    return items.sort(function(x, y) {
-      if (!/^\d+$/.test(x) || !/^\d+$/.test(y)) {
-        if (x > y) {
-          return 1;
-        } else if (y > x) {
-          return -1;
-        }
-        return 1;
-      }
-      if (parseInt(x) > parseInt(y)) {
-        return 1;
-      } else if (parseInt(y) > parseInt(x)) {
-        return -1;
-      }
-      return 1;
-    });
-  } else if (sort == 'Toughness') {
-    var items = [];
-    cube.forEach(function(card, index) {
-      if (card.details.toughness) {
-        if (!items.includes(card.details.toughness)) {
-          items.push(card.details.toughness);
-        }
-      }
-    });
-    return items.sort(function(x, y) {
-      if (!/^\d+$/.test(x) || !/^\d+$/.test(y)) {
-        if (x > y) {
-          return 1;
-        } else if (y > x) {
-          return -1;
-        }
-        return 1;
-      }
-      if (parseInt(x) > parseInt(y)) {
-        return 1;
-      } else if (parseInt(y) > parseInt(x)) {
-        return -1;
-      }
-      return 1;
-    });
-  } else if (sort == 'Loyalty') {
-    var items = [];
-    cube.forEach(function(card, index) {
-      if (card.details.loyalty) {
-        if (!items.includes(card.details.loyalty)) {
-          items.push(card.details.loyalty);
-        }
-      }
-    });
-    return items.sort(function(x, y) {
-      if (!/^\d+$/.test(x) || !/^\d+$/.test(y)) {
-        if (x > y) {
-          return 1;
-        } else if (y > x) {
-          return -1;
-        }
-        return 1;
-      }
-      if (parseInt(x) > parseInt(y)) {
-        return 1;
-      } else if (parseInt(y) > parseInt(x)) {
-        return -1;
-      }
-      return 1;
-    });
-  } else if (sort == 'Manacost Type') {
-    return ['Gold', 'Hybrid', 'Phyrexian'];
-  } else if (sort == 'CNC') {
-    return ['Creature', 'Non-Creature'];
-  } else if (sort == 'Price' || sort == 'Price Foil') {
-    var labels = [];
-    for (i = 0; i <= price_buckets.length; i++) {
-      labels.push(price_bucket_label(i));
-    }
-    labels.push('No Price Available');
-    return labels;
-  } else if (sort == 'Unsorted') {
-    return ['All'];
-  }
-}
-
-function sortIntoGroups(cards, sort) {
-  var groups = {};
-  var labels = getLabels(sort);
-  labels.forEach(function(label, index) {
-    group = [];
-
-    cards.forEach(function(card, cardindex) {
-      if (cardIsLabel(card, label, sort)) {
-        group.push(card);
-      }
-    });
-
-    if (group.length > 0) {
-      groups[label] = group;
-    }
-  });
-  return groups;
-}
