+++ src/components/TableView.js
@@ -2,49 +2,39 @@ import React, { useEffect } from 'react';
 
 import { Row, Col } from 'reactstrap';
 
+import { countGroup, sortDeep } from '../util/Sort';
+
 import AutocardListGroup from './AutocardListGroup';
 import SortContext from './SortContext';
 
 const TableViewRaw = ({ cards, primary, secondary, tertiary, changeSort, ...props }) => {
-  let columns = sortIntoGroups(cards, primary);
-  let columnCounts = {};
-  for (let columnLabel of Object.keys(columns)) {
-    columnCounts[columnLabel] = columns[columnLabel].length;
-    columns[columnLabel] = sortIntoGroups(columns[columnLabel], secondary);
-  }
+  const sorted = sortDeep(cards, primary, secondary);
 
   return (
     <div className="table-view-container">
       <Row className="table-view" {...props}>
-        {getLabels(primary)
-          .filter((columnLabel) => columns[columnLabel])
-          .map((columnLabel) => {
-            let column = columns[columnLabel];
-            return (
-              <Col
-                key={columnLabel}
-                xs="6"
-                md="3"
-                lg="auto"
-                className="mt-3 table-col"
-                style={{ width: `${100 / Math.min(Object.keys(columns).length, 8)}%` }}
-              >
-                <h6 className="text-center">
-                  {columnLabel}
-                  <br />({columnCounts[columnLabel]})
-                </h6>
-                {getLabels(secondary)
-                  .filter((label) => column[label])
-                  .map((label) => (
-                    <AutocardListGroup
-                      key={label}
-                      heading={`${label} (${column[label].length})`}
-                      cards={column[label]}
-                    />
-                  ))}
-              </Col>
-            );
-          })}
+        {sorted.map(([columnLabel, column]) => (
+          <Col
+            key={columnLabel}
+            xs="6"
+            md="3"
+            lg="auto"
+            className="mt-3 table-col"
+            style={{ width: `${100 / Math.min(sorted.length, 8)}%` }}
+          >
+            <h6 className="text-center">
+              {columnLabel}
+              <br />({countGroup(column)})
+            </h6>
+            {column.map(([label, row]) => (
+              <AutocardListGroup
+                key={label}
+                heading={`${label} (${countGroup(row)})`}
+                cards={row}
+              />
+            ))}
+          </Col>
+        ))}
       </Row>
     </div>
   );
