+++ src/util/Sort.js
@@ -1,4 +1,43 @@
-var price_buckets = [0.25, 0.5, 1, 2, 3, 4, 5, 7, 10, 15, 20, 25, 30, 40, 50, 75, 100];
+import { alphaCompare } from './Util';
+
+function ISODateToYYYYMMDD(dateString) {
+  const locale = 'en-US';
+
+  if (dateString === undefined) {
+    return undefined;
+  }
+
+  return new Date(dateString).toLocaleDateString(locale);
+}
+
+function GetColorIdentity(colors) {
+  if (colors.length == 0) {
+    return 'Colorless';
+  } else if (colors.length > 1) {
+    return 'Multicolored';
+  } else if (colors.length == 1) {
+    switch (colors[0]) {
+      case 'W':
+        return 'White';
+        break;
+      case 'U':
+        return 'Blue';
+        break;
+      case 'B':
+        return 'Black';
+        break;
+      case 'R':
+        return 'Red';
+        break;
+      case 'G':
+        return 'Green';
+        break;
+      case 'C':
+        return 'Colorless';
+        break;
+    }
+  }
+}
 
 function GetColorCategory(type, colors) {
   if (type.toLowerCase().includes('land')) {
@@ -31,6 +70,276 @@ function GetColorCategory(type, colors) {
   }
 }
 
+export function getSorts() {
+  return [
+    'Artist',
+    'CMC',
+    'Color Category',
+    'Color Count',
+    'Color Identity',
+    'Color',
+    'Date Added',
+    'Finish',
+    'Guilds',
+    'Legality',
+    'Loyalty',
+    'Manacost Type',
+    'Power',
+    'Price',
+    'Price Foil',
+    'Rarity',
+    'Set',
+    'Shards / Wedges',
+    'Status',
+    'Subtype',
+    'Supertype',
+    'Tags',
+    'Toughness',
+    'Type',
+    'Types-Multicolor',
+    'Unsorted',
+  ];
+}
+
+export function getLabels(cube, sort) {
+  if (sort == 'Color Category') {
+    return ['White', 'Blue', 'Black', 'Red', 'Green', 'Multicolored', 'Colorless', 'Lands'];
+  } else if (sort == 'Color Identity') {
+    return ['White', 'Blue', 'Black', 'Red', 'Green', 'Multicolored', 'Colorless'];
+  } else if (sort == 'CMC') {
+    return ['0', '1', '2', '3', '4', '5', '6', '7', '8+'];
+  } else if (sort == 'CMC2') {
+    return ['0-1', '2', '3', '4', '5', '6', '7+'];
+  } else if (sort == 'CMC-Full') {
+    // All CMCs from 0-16, with halves included, plus Gleemax at 1,000,000.
+    return Array.from(Array(33).keys())
+      .map((x) => (x / 2).toString())
+      .concat(['1000000']);
+  } else if (sort == 'Color') {
+    return ['White', 'Blue', 'Black', 'Red', 'Green', 'Colorless'];
+  } else if (sort == 'Type') {
+    return [
+      'Creature',
+      'Planeswalker',
+      'Instant',
+      'Sorcery',
+      'Artifact',
+      'Enchantment',
+      'Conspiracy',
+      'Contraption',
+      'Phenomenon',
+      'Plane',
+      'Scheme',
+      'Vanguard',
+      'Land',
+      'Other',
+    ];
+  } else if (sort == 'Supertype') {
+    return ['Snow', 'Legendary', 'Tribal', 'Basic', 'Elite', 'Host', 'Ongoing', 'World'];
+  } else if (sort == 'Tags') {
+    var tags = [];
+    cube.forEach(function(card, index) {
+      card.tags.forEach(function(tag, index2) {
+        if (tag.length > 0 && !tags.includes(tag)) {
+          tags.push(tag);
+        }
+      });
+    });
+    return tags.sort();
+  } else if (sort == 'Date Added') {
+    var days = [],
+      formattedDay;
+    cube.forEach(function(card, index) {
+      formattedDay = ISODateToYYYYMMDD(card.addedTmsp);
+      if (formattedDay === undefined) {
+        formattedDay = 'unknown';
+      }
+      if (!days.includes(formattedDay)) {
+        days.push(formattedDay);
+      }
+    });
+    return days.sort();
+  } else if (sort == 'Status') {
+    return ['Not Owned', 'Ordered', 'Owned', 'Premium Owned'];
+  } else if (sort == 'Finish') {
+    return ['Non-foil', 'Foil'];
+  } else if (sort == 'Guilds') {
+    return ['Azorius', 'Dimir', 'Rakdos', 'Gruul', 'Selesnya', 'Orzhov', 'Golgari', 'Simic', 'Izzet', 'Boros'];
+  } else if (sort == 'Shards / Wedges') {
+    return ['Bant', 'Esper', 'Grixis', 'Jund', 'Naya', 'Abzan', 'Jeskai', 'Sultai', 'Mardu', 'Temur'];
+  } else if (sort == 'Color Count') {
+    return ['0', '1', '2', '3', '4', '5'];
+  } else if (sort == 'Set') {
+    var sets = [];
+    cube.forEach(function(card, index) {
+      if (!sets.includes(card.details.set.toUpperCase())) {
+        sets.push(card.details.set.toUpperCase());
+      }
+    });
+    return sets.sort();
+  } else if (sort == 'Artist') {
+    var artists = [];
+    cube.forEach(function(card, index) {
+      if (!artists.includes(card.details.artist)) {
+        artists.push(card.details.artist);
+      }
+    });
+    return artists.sort();
+  } else if (sort == 'Rarity') {
+    return ['Common', 'Uncommon', 'Rare', 'Mythic'];
+  } else if (sort == 'Unsorted') {
+    return ['All'];
+  } else if (sort == 'Subtype') {
+    var types = [];
+    cube.forEach(function(card, index) {
+      if (card.type_line.includes('—')) {
+        var subtypes = card.type_line.substr(card.type_line.indexOf('—') + 1).split(' ');
+        subtypes.forEach(function(subtype, index) {
+          if (!types.includes(subtype.trim()) && subtype.trim().length > 0) {
+            types.push(subtype.trim());
+          }
+        });
+      }
+    });
+    return types.sort();
+  } else if (sort == 'Types-Multicolor') {
+    return [
+      'Creature',
+      'Planeswalker',
+      'Instant',
+      'Sorcery',
+      'Artifact',
+      'Enchantment',
+      'Conspiracy',
+      'Contraption',
+      'Phenomenon',
+      'Plane',
+      'Scheme',
+      'Vanguard',
+      'Azorius',
+      'Dimir',
+      'Rakdos',
+      'Gruul',
+      'Selesnya',
+      'Orzhov',
+      'Golgari',
+      'Simic',
+      'Izzet',
+      'Boros',
+      'Bant',
+      'Esper',
+      'Grixis',
+      'Jund',
+      'Naya',
+      'Abzan',
+      'Jeskai',
+      'Sultai',
+      'Mardu',
+      'Temur',
+      'Non-White',
+      'Non-Blue',
+      'Non-Black',
+      'Non-Red',
+      'Non-Green',
+      'Five Color',
+      'Land',
+      'Other',
+    ];
+  } else if (sort == 'Legality') {
+    return ['Standard', 'Modern', 'Legacy', 'Vintage', 'Pauper'];
+  } else if (sort == 'Power') {
+    var items = [];
+    cube.forEach(function(card, index) {
+      if (card.details.power) {
+        if (!items.includes(card.details.power)) {
+          items.push(card.details.power);
+        }
+      }
+    });
+    return items.sort(function(x, y) {
+      if (!/^\d+$/.test(x) || !/^\d+$/.test(y)) {
+        if (x > y) {
+          return 1;
+        } else if (y > x) {
+          return -1;
+        }
+        return 1;
+      }
+      if (parseInt(x) > parseInt(y)) {
+        return 1;
+      } else if (parseInt(y) > parseInt(x)) {
+        return -1;
+      }
+      return 1;
+    });
+  } else if (sort == 'Toughness') {
+    var items = [];
+    cube.forEach(function(card, index) {
+      if (card.details.toughness) {
+        if (!items.includes(card.details.toughness)) {
+          items.push(card.details.toughness);
+        }
+      }
+    });
+    return items.sort(function(x, y) {
+      if (!/^\d+$/.test(x) || !/^\d+$/.test(y)) {
+        if (x > y) {
+          return 1;
+        } else if (y > x) {
+          return -1;
+        }
+        return 1;
+      }
+      if (parseInt(x) > parseInt(y)) {
+        return 1;
+      } else if (parseInt(y) > parseInt(x)) {
+        return -1;
+      }
+      return 1;
+    });
+  } else if (sort == 'Loyalty') {
+    var items = [];
+    cube.forEach(function(card, index) {
+      if (card.details.loyalty) {
+        if (!items.includes(card.details.loyalty)) {
+          items.push(card.details.loyalty);
+        }
+      }
+    });
+    return items.sort(function(x, y) {
+      if (!/^\d+$/.test(x) || !/^\d+$/.test(y)) {
+        if (x > y) {
+          return 1;
+        } else if (y > x) {
+          return -1;
+        }
+        return 1;
+      }
+      if (parseInt(x) > parseInt(y)) {
+        return 1;
+      } else if (parseInt(y) > parseInt(x)) {
+        return -1;
+      }
+      return 1;
+    });
+  } else if (sort == 'Manacost Type') {
+    return ['Gold', 'Hybrid', 'Phyrexian'];
+  } else if (sort == 'CNC') {
+    return ['Creature', 'Non-Creature'];
+  } else if (sort == 'Price' || sort == 'Price Foil') {
+    var labels = [];
+    for (i = 0; i <= price_buckets.length; i++) {
+      labels.push(price_bucket_label(i));
+    }
+    labels.push('No Price Available');
+    return labels;
+  } else if (sort == 'Unsorted') {
+    return ['All'];
+  }
+}
+
+var price_buckets = [0.25, 0.5, 1, 2, 3, 4, 5, 7, 10, 15, 20, 25, 30, 40, 50, 75, 100];
+
 //returns the price bucket label at the index designating the upper bound
 //at index == 0, returns < lowest
 //at index == length, returs >= highest
@@ -53,7 +362,7 @@ function cmcToNumber(card) {
   }
 }
 
-function cardIsLabel(card, label, sort) {
+export function cardIsLabel(card, label, sort) {
   if (sort == 'Color Category') {
     return GetColorCategory(card.type_line, card.colors) == label;
   } else if (sort == 'Color Identity') {
@@ -337,12 +646,37 @@ function cardIsLabel(card, label, sort) {
   }
 }
 
-try {
-  module.exports = {
-    cardIsLabel: cardIsLabel,
-    filterCard: filterCard,
-    price_buckets: price_buckets,
-  };
-} catch (err) {
-  //probably running client side, ignore
+export function sortIntoGroups(cards, sort) {
+  const groups = {};
+  const labels = getLabels(cards, sort);
+  for (const label of labels) {
+    const group = cards.filter(card => cardIsLabel(card, label, sort));
+
+    if (group.length > 0) {
+      groups[label] = group;
+    }
+  }
+  return groups;
+}
+
+function sortGroupsOrdered(cards, sort) {
+  return getLabels(cards, sort).map(label => [label, cards.filter(card => cardIsLabel(card, label, sort))]).filter(([label, group]) => group.length > 0);
+}
+
+export function sortDeep(cards, ...sorts) {
+  if (sorts.length === 0) {
+    return cards.sort(alphaCompare);
+  } else {
+    const [first, ...rest] = sorts;
+    return getLabels(cards, first).map(label => [label, sortDeep(cards.filter(card => cardIsLabel(card, label, first)), ...rest)]).filter(([label, group]) => group.length > 0);
+  }
+}
+
+export function countGroup(group) {
+  if (Array.isArray(group[0])) {
+    const counts = group.map(([label, group2]) => countGroup(group2));
+    return counts.reduce((a, b) => a + b, 0);
+  } else {
+    return group.length;
+  }
 }
