+++ views/cube/bulk_upload.pug
@@ -37,5 +37,6 @@ block content
             input#changelistBlog(type='hidden', name='blog')
             input.btn.mr-sm-2.btn-success#saveChangesButton(type='button', value='Save Changes')
 
+block scripts
   script(src='/js/autocomplete.js')
   script(src='/js/bulkconfirm.js')
