+++ views/cube/cube_blog.pug
@@ -57,7 +57,7 @@ block cube_content
               input.btn.btn-danger.delete-blog(type='submit',value='Delete', data-id='', id="delete")
               button.btn.btn-secondary(type='button', data-dismiss='modal') Cancel
 
+block scripts
   include ../react
-  script(src='/js/csrf.js')
   script(src='/js/blog.js')
   script(src='/js/cube_blog.bundle.js')
