+++ views/cube/cube_compare.pug
@@ -32,9 +32,7 @@ block content
   //Cube list
   #react-root
 
+block scripts
   include ../react
-  script(src='/js/csrf.js')
-  script(src='/js/sortfilter.js')
-  script(src='/js/editcube.js')
   script(src='/js/autocard.js')
   script(src='/js/cube_compare.bundle.js')
