+++ views/cube/cube_deck.pug
@@ -7,6 +7,7 @@ block cube_content
   include ../dynamic_flash
   #react-root
 
+block scripts
   script(type='text/javascript').
     var reactProps = {
       oldFormat: !{oldformat},
