+++ views/cube/cube_deckbuilder.pug
@@ -7,6 +7,7 @@ block cube_content
   include ../dynamic_flash
   #react-root
 
+block scripts
   script(type='text/javascript').
     var reactProps = {
       initialDeck: !{deck_raw},
