+++ views/cube/cube_draft.pug
@@ -8,6 +8,8 @@ block cube_content
   #react-root
   script(type='text/javascript').
     var initialDraft = !{draft_raw};
+
+block scripts
   include ../react
   script(src='/js/autocard.js')
   script(src='/js/cube_draft.bundle.js')
