+++ views/cube/cube_overview.pug
@@ -49,6 +49,8 @@ block cube_content
             .modal-footer
               input.btn.btn-danger.delete-cube(type='submit',value='Delete', data-id=cube_id, disabled="", id="delete")
               button.btn.btn-secondary(type='button', data-dismiss='modal') Cancel
+
+block scripts
   include ../react
   script(src='/js/csrf.js')
   script(src='/js/blog.js')
