+++ views/dashboard.pug
@@ -10,5 +10,6 @@ block content
   input#decksData(type='hidden', value=decks)
   input#userid(type='hidden', value=user._id)
 
+block scripts
   include react
   script(src='/js/dashboard.bundle.js')
