+++ views/index.pug
@@ -142,4 +142,6 @@ block content
 							p.card-text !{devblog.html}
 						p
 							a(href='/dev/blog') See more..
+
+block scripts
 	script(src='/js/index.js')
